spostato le funzioni Api su un controller specifico
