+++ app/Http/Controllers/AuthorizedUsers/MarksController.php
@@ -22,44 +22,6 @@ class MarksController extends CommonController
      */
     public function index(Request $request)
     {
-        /* $userId = Auth::id();
-        $user = Auth::user();
-        $user_role = $user->role;
-        $students = [];
-        $classes = [];
-        $page = 'Voti';
-
-        $teacher = Teacher::where('user_id', $userId)->first();
-
-        if ($teacher) {
-            $classes = $teacher->classes;
-
-            $grades = GradesStudentRegister::where('teacher_id', $teacher->id)
-                                            ->where('type', 'mark')
-                                            ->get();
-
-            if ($classes->isNotEmpty()) {
-                $selectedClassId = $request->input('selected_class');
-
-                if ($selectedClassId) {
-                    $selectedClass = $classes->where('id', $selectedClassId)->first();
-
-                    if ($selectedClass) {
-                        $students = $selectedClass->students;
-                    } else {
-                        return view('teacher.presents', compact('students', 'classes', 'user_role', 'page'))->withErrors(['message' => 'Invalid selected class.']);
-                    }
-                } else {
-                    $students = $classes->first()->students;
-                }
-
-                return view('teacher.presents', compact('students', 'classes', 'user_role', 'page', 'grades'));
-            } else {
-                return view('teacher.presents', compact('students', 'classes', 'user_role', 'page'))->withErrors(['message' => 'No classes found for the teacher.']);
-            }
-        } else {
-            return view('teacher.presents', compact('students', 'classes', 'user_role', 'page'))->withErrors(['message' => 'Teacher not found.']);
-        } */
         return $this->commonIndex($request, 'Voti');
     }
 
@@ -106,7 +68,7 @@ class MarksController extends CommonController
             $data = $validatedData['date'];
             $formattedDate = date('Y-m-d', strtotime($data));
 
-            \Log::info($formattedDate);
+            Log::info($formattedDate);
 
             $mark = GradesStudentRegister::create([
                 'student_id' => $validatedData['student'],
@@ -169,68 +131,6 @@ class MarksController extends CommonController
 
         // Ritorna una risposta di successo
         return response()->json(['message' => 'Voto eliminato con successo.']);
-    }
-
-    /**
-     * Recupera i voti per quella classe
-     */
-
-    public function getGrades(Request $request)
-    {
-        // Verifica se è stata fornita una query string per la materia
-        if ($request->has('subject')) {
-            // Recupera i voti in base alla materia selezionata
-            $subjectId = $request->input('subject');
-            $grades = GradesStudentRegister::where('type', 'mark')
-                                            ->where('subject_id', $subjectId)
-                                            ->get();
-        } else {
-            $grades = [];
-        }
-    
-        return response()->json($grades);
-    }
-
-    /**
-     * Recupera le opzioni voti
-     */
-
-    public function getGradesOption()
-    {
-        $gradeOptions = GradeOption::all();
-        return response()->json($gradeOptions);
-    }
-
-    /**
-     * Recupera le opzioni materie
-     */
-
-    public function getSubjectsOption()
-    {
-        $subjectOption = Subject::all();
-        return response()->json($subjectOption);
-    }
-
-    /**
-     * Recupera gli studenti di quella classe
-     */
-
-    public function getStudentsByClass(Request $request)
-    {
-        //Log::info($request);
-        if ($request->has('class')) {
-            $classeId = $request->input('class');
-            $classe = Classe::find($classeId);
-
-            if ($classe) {
-                $students = Student::where('class_id', $classeId)->get();
-                return response()->json($students);
-            } else {
-                return response()->json(['message' => 'Classe non trovata'], 404);
-            }
-        } else {
-            return response()->json(['message' => 'Parametro "class" mancante'], 400);
-        }
     }
 
 }
