+++ app/Http/Controllers/AuthorizedUsers/PresenceController.php
@@ -27,88 +27,6 @@ class PresenceController extends CommonController
      */
     public function index(Request $request)
     {
-        /* $user = Auth::user();
-        if (!$user){
-            abort(401, 'Unauthorized');
-        }
-        $userId = $user->id;
-        $user_role = $user->role;
-        $students = [];
-        $classes = [];
-        $timetable = [];
-        $res = [];
-        $page = 'Presenze';
-        $current_date = date("Y-m-d");
-        $current_day = strftime('%A');
-        if($request->input('current_date') !== null)
-        {
-           
-            $rules = [
-                'current_date' => ['regex:/^\d{1,2}\s(?:January|February|March|April|May|June|July|August|September|October|November|December)\s\d{4}$/i'],
-            ];
-            $messages = [
-                'current_date.regex' => 'Field :attribute MUST BE in this format: day, month in letter in English INGLESE and year.',
-            ];
-            $validator = Validator::make($request->all(), $rules, $messages);
-            if ($validator->fails()) {
-                return back()->withErrors($validator)->withInput();
-            }
-            $current_date = $request->input('current_date');
-            $current_date = \DateTime::createFromFormat('j F Y', $current_date)->format('Y-m-d');
-            $current_day = date("l", strtotime($current_date));
-        }
-
-        $teacher = Teacher::where('user_id', $userId)->first();
-
-        if ($teacher) {
-            $classes = $teacher->classes;
-
-            if ($classes->isNotEmpty()) {
-                
-                $rules = [
-                    'selected_class' => 'integer',
-                ];
-                $messages = [
-                    'selected_class.integer' => 'Field :attribute MUST BE an id, an integer number',
-                ];
-                $validator = Validator::make($request->all(), $rules, $messages);
-                if ($validator->fails()) {
-                    return back()->withErrors($validator)->withInput();
-                }
-                $selectedClassId = $request->input('selected_class');
-                
-                //Log::info($selectedClassId);    
-
-                if ($selectedClassId) {
-                    $selectedClass = $classes->where('id', $selectedClassId)->first();                
-                    if ($selectedClass) {
-                        $students = $selectedClass->students;
-                        $timetable = $selectedClass->calendar;
-                        //costruzione matrice per la griglia delle presenze
-                        $res = $this->buildPresenceGrid($students,$timetable,$current_day,$current_date);
-                        if ($res === false){
-                            $res = [];
-                            return view('teacher.presents', compact('students', 'classes', 'user_role', 'page', 'timetable', 'res', 'current_date','current_day'))->withErrors(['message' => 'Invalid Timetable']);                            
-                        }
-                    } else {
-                        return view('teacher.presents', compact('students', 'classes', 'user_role', 'page', 'timetable', 'res', 'current_date','current_day'))->withErrors(['message' => 'Invalid selected class.']);
-                    }
-                } else {
-                    $students = $classes->first()->students;
-                    $timetable = $classes->first()->calendar;
-                    //costruzione matrice per la griglia delle presenze
-                    $res = $this->buildPresenceGrid($students,$timetable,$current_day,$current_date);
-                    if ($res === false){
-                        return view('teacher.presents', compact('students', 'classes', 'user_role', 'page', 'timetable', 'res', 'current_date','current_day'))->withErrors(['message' => 'Invalid Timetable']);                            
-                    }
-                }
-                return view('teacher.presents', compact('students', 'classes', 'user_role', 'page', 'timetable', 'res','current_date','current_day'));
-            } else {
-                return view('teacher.presents', compact('students', 'classes', 'user_role', 'page', 'timetable', 'res','current_date','current_day'))->withErrors(['message' => 'No classes found for the teacher.']);
-            }
-        } else {
-            return view('teacher.presents', compact('students', 'classes', 'user_role', 'page', 'timetable', 'res','current_date','current_day'))->withErrors(['message' => 'Teacher not found.']);
-        } */
         return $this->commonIndex($request, 'Presenze');
     }
 
@@ -276,76 +194,4 @@ class PresenceController extends CommonController
         return $responseData;
     }
 
-    public function getTimetable(string $id, string $dateParam)
-    {
-        //Log::info($id);
-        $dayOfWeek = date("l", strtotime($dateParam));
-        //Log::info($dayOfWeek);
-        try{
-            $selectedClass = Classe::findOrFail($id);
-            //Log::info($selectedClass->id);    
-        } catch (\Illuminate\Database\Eloquent\ModelNotFoundException $e){
-            Log::info($e);
-        }
-             
-        if ($selectedClass) {
-            $timetable = $selectedClass->calendar();
-            //Log::info($timetable->pluck('time_start')->implode(', '));
-            $filteredTimetable = $timetable->where('day_of_week', $dayOfWeek)->get();
-            //Log::info(empty($filteredTimetable));
-            if($filteredTimetable->isEmpty()){
-                return response()->json(['message' => 'La classe non ha un orario associato']);
-            }
-            //costruzione matrice per la griglia delle presenze
-            //$res = $this->buildPresenceGrid($students,$timetable,$current_day,$current_date);
-            
-            $matrice = $this->getPresences($filteredTimetable ,$selectedClass, $dateParam);
-            /* if(empty($matrice)){
-                return response()->json(['message' => 'Gli studenti non hanno presenze/assenze da mostrare']);
-            } */
-            $resp = [
-                'timetable' => $filteredTimetable,
-                'presences' => $matrice
-            ];
-            //Log::info($resp);
-            return response()->json($resp);
-        } else {
-            return response()->json(['message' => 'Orario non trovato'], 404);
-        }
-    }
-
-    public function getPresences($timetable, $selectedClass, $current_date){
-
-        $students = $selectedClass->students;
-        $current_date = \DateTime::createFromFormat('j F Y', $current_date)->format('Y-m-d');
-        $res = [];
-        foreach($students as $student){
-            $dayPresence = json_decode($student->presences, true);
-            //Filtra gli elementi dell'array in base al giorno corrente
-            $dayPresences = array_filter($dayPresence, function($item) use ($current_date) {
-                return $item['data'] === $current_date;
-            });
-            
-            $values = array_values($dayPresences);
-            foreach($timetable as $t){
-                $trovato = false;
-                foreach($values as $v){
-                    if($t['time_start'] ==  $v['hour']){
-                        $res[$student->id][] = [$v['presence'], $v['id']];  
-                        //$res[$student->id][] = $v['presence'];
-                        $trovato = true;
-                    } else {
-                        if ($trovato){
-                            break;
-                        }   
-                    }  
-                }
-                if(!$trovato){
-                    $res[$student->id][] = ['',''];
-                }
-            } 
-        }
-        //Log::info($res);
-        return $res;
-    }
 }
